+++ src/main/java/bowling/domain/player/PlayerName.java
@@ -1,4 +1,4 @@
-package bowling.domain;
+package bowling.domain.player;
 
 import bowling.util.StringUtil;
